+++ source/LDL/Shared/Graphics/Math/Isometric.cpp
@@ -9,8 +9,8 @@ using namespace LDL::Math;
 
 const Vec2u& Isometric::CartesianToIsometric(const Vec2u& pt)
 {
-	size_t x = pt.x - pt.y;
-	size_t y = (pt.x + pt.y) / 2;
+	uint32_t x = pt.x - pt.y;
+	uint32_t y = (pt.x + pt.y) / 2;
 
 	_result = Vec2u(x, y);
 
@@ -19,8 +19,8 @@ const Vec2u& Isometric::CartesianToIsometric(const Vec2u& pt)
 
 const Vec2u& Isometric::IsometricToCartesian(const Vec2u& pt)
 {
-	size_t x = (2 * pt.y + pt.x) / 2;
-	size_t y = (2 * pt.y - pt.x) / 2;
+	uint32_t x = (2 * pt.y + pt.x) / 2;
+	uint32_t y = (2 * pt.y - pt.x) / 2;
 
 	_result = Vec2u(x, y);
